requirements.txt
openai 
Pyrogram 
pydub
python-dotenv